+++ src/python/dxpy/nextflow/nextflow_builder.py
@@ -138,16 +138,20 @@ def prepare_custom_inputs(schema_file="./nextflow_schema.json"):
             dx_input = {}
             dx_input["name"] = property_key
             dx_input["title"] = dx_input['name']
+            if "default" in property:
+                dx_input["help"] = "Default value:{}\n".format(property.get("default"))
             if "help_text" in property:
-                dx_input["help"] = property.get('help_text')
+                dx_input["help"] = dx_input.get("help") + property.get('help_text')
             dx_input["hidden"] = property.get('hidden', False)
             dx_input["class"] = get_dx_type(property.get("type"), property.get("format"))
             dx_input["optional"] = True
             if property_key not in required_inputs:
                 dx_input["help"] = "(Optional) {}".format(dx_input.get("help", ""))
+                dx_input["group"] = "Optional Nextflow inputs"
                 inputs.append(dx_input)
             else:
                 dx_input["help"] = "(Required) {}".format(dx_input.get("help", ""))
+                dx_input["group"] = "Required Nextflow inputs"
                 inputs.insert(0, dx_input)
 
     return inputs